+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/history/summary/index.svelte
@@ -13,7 +13,6 @@
     return {
       props: {
         items,
-        category,
       },
       stuff: {
         matchingEvents: events,
@@ -29,13 +28,10 @@
   import Pagination from '$lib/components/pagination.svelte';
   import LoadingRow from '$lib/components/loading-row.svelte';
 
-  import FilterSelect from '$lib/components/select/filter-select.svelte';
-  import Option from '$lib/components/select/option.svelte';
   import EventSummaryTable from '$lib/components/event/event-summary-table.svelte';
   import EventSummaryRow from '$lib/components/event/event-summary-row.svelte';
 
   export let items: HistoryEventWithId[];
-  export let category: EventTypeCategory;
 
   const startingIndex = items.findIndex(
     ({ id }) => $page.params.eventId === id,
@@ -46,17 +42,6 @@
   <LoadingRow />
 {:then items}
   {#if items.length}
-    <div class="flex items-center gap-4">
-      <FilterSelect parameter="category" bind:value={category}>
-        <Option value={undefined}>All</Option>
-        <Option value="activity">Activity</Option>
-        <Option value="command">Command</Option>
-        <Option value="signal">Signal</Option>
-        <Option value="timer">Timer</Option>
-        <Option value="child-workflow">Child Workflow</Option>
-        <Option value="workflow">Workflow</Option>
-      </FilterSelect>
-    </div>
     <Pagination {items} {startingIndex} let:visibleItems>
       <EventSummaryTable>
         {#each visibleItems as event (event.id)}
@@ -65,6 +50,10 @@
       </EventSummaryTable>
     </Pagination>
   {:else}
-    <EmptyState title={'No Events Found'} />
+    <Pagination {items} {startingIndex} let:visibleItems>
+      <EventSummaryTable>
+        <EmptyState title={'No Events Found'} />
+      </EventSummaryTable>
+  </Pagination>
   {/if}
 {/await}
